clean up and standardize instructions
what
===

cleans up and standardizes instructions to make them clearer. changes
include:
- rewrite instructions with the assumption of cloud-based IDE (hackerrank)
- remove 'bring your own implementation' instructions
- combine on-site and remote instructions

why
===

> - rewrite instructions with the assumption of cloud-based IDE (hackerrank)

afaik, we use hackerrank for all interviews now, so let's remove any
expectations to setup the application locally as it'll already be setup
in a cloud-based IDE instead. lmk if this is not the case.

> - remove 'bring your own implementation' instructions

this is a rarely used edge case. let's standardize on the existing ruby,
python, and js implementations that are ready to spin up in the cloud.

> - combine on-site and remote instructions

afaik, all coding interviews are 100% remote now, so let's remove on-site
instructions.